+++ packages/errors-nest/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.7.0
+
+### Minor Changes
+
+- Setup Domain Event Basis
+
 ## 0.6.6
 
 ### Patch Changes